requirements.txt
Django>=4.2.0
django-allauth>=0.57.0
django-cors-headers>=4.3.0
django-extensions>=4.1.0
djangorestframework>=3.14.0
dj-rest-auth>=5.0.0
drf-yasg>=1.21.7
mysqlclient>=2.2.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
Werkzeug>=3.0.1
pyOpenSSL>=24.0.0 